Add ApplicationDbContext class to interact with database
This commit adds an ApplicationDbContext class to the BlazorToDo.Data namespace. The ApplicationDbContext class inherits from the DbContext class provided by the Microsoft.EntityFrameworkCore namespace, and is used to interact with the database in .NET applications. The class includes a DbSet property called TodoList, which represents a collection of Todo objects in the database.