Add authentication mechanism
This commit adds authentication to To Do Blazor. Users will be able to register using an email and password, currently without receiving a confirmation email, and access the password-protected pages. Users will only be able to access pages restricted by the Authorize attribute after logging in. In addition, the structure of the sidebar will be adjusted accordingly when the user logs in or out.